crm-25: making sure we dont delete the logged user

+++ app/Livewire/Admin/Users/Delete.php
@@ -38,6 +38,13 @@ class Delete extends Component
     {
         $this->validate();
 
+        if ($this->user->is(auth()->user())) {
+
+            $this->addError('confirmation', 'You cannot delete your own account.');
+
+            return;
+        }
+
         $this->user->delete();
 
         $this->user->notify(new UserDeletedNotificaion());
